workspace: Remove WindowClone's from _windows when destroyed
A WindowClone might be destroyed earlier than its MetaWindow counterpart
as its WindowActor could be destroyed earlier, thus when this happens
it's safer to remove the clone from the windows list, without waiting
the workspace to request to do so.

WindowClone now emits a 'destroy' signals earlier enough and this now
triggers a _doRemoveWindow on Workspace which will lead to the proper
cleanup; keeping track of the signal connections, in order to avoid
callback loops (not really harmful in this case, but good practice).

https://bugzilla.gnome.org/show_bug.cgi?id=791233